+++ crates/icu/src/interface/response.rs
@@ -1,6 +1,5 @@
 use crate::{
     Error,
-    canister::CanisterRegistry,
     ic::api::{canister_cycle_balance, msg_caller},
     interface::{
         InterfaceError,
@@ -8,7 +7,8 @@ use crate::{
         request::{CreateCanisterRequest, CyclesRequest, Request, UpgradeCanisterRequest},
         state::{StateBundle, cascade, update_canister},
     },
-    memory::{CanisterState, SubnetIndex},
+    memory::{CanisterState, SubnetDirectory, SubnetRegistry, SubnetRegistryEntry},
+    state::canister::CanisterRegistry,
 };
 use candid::{CandidType, Principal};
 use serde::Deserialize;
@@ -53,6 +53,8 @@ pub struct CyclesResponse {
 
 // response
 pub async fn response(req: Request) -> Result<Response, Error> {
+    assert!(CanisterState::is_root(), "only root can run this code");
+
     match req {
         Request::CreateCanister(req) => create_canister_response(&req).await,
         Request::UpgradeCanister(req) => upgrade_canister_response(&req).await,
@@ -63,30 +65,33 @@ pub async fn response(req: Request) -> Result<Response, Error> {
 // create_canister_response
 async fn create_canister_response(req: &CreateCanisterRequest) -> Result<Response, Error> {
     let canister = CanisterRegistry::try_get(&req.kind)?;
-    let is_root = CanisterState::is_root();
-    let is_indexable = canister.attributes.is_indexable();
+    let created_under_root = req.parents.len() <= 1;
+    let uses_directory = canister.attributes.uses_directory();
 
-    // indexable canisters have to be on root
+    // canisters have to be on root
     // and subnet_index has to be able to accept them
-    if is_indexable {
-        if is_root {
-            SubnetIndex::can_insert(&canister)?;
+    if uses_directory {
+        if created_under_root {
+            SubnetDirectory::can_insert(&canister)?;
         } else {
             return Err(InterfaceError::CannotCreateIndexable)?;
         }
     }
+    let kind = req.kind.clone();
 
-    let new_canister_pid = ic_create_canister(
-        &req.kind,
-        canister.wasm,
-        &req.parents,
-        req.extra_arg.clone(),
-    )
-    .await?;
+    let new_canister_pid =
+        ic_create_canister(&kind, canister.wasm, &req.parents, req.extra_arg.clone()).await?;
+
+    // add to SubnetRegistry
+    SubnetRegistry::insert(SubnetRegistryEntry {
+        canister_pid: new_canister_pid,
+        kind,
+        parent_pid: req.parents.last().map(|p| p.principal),
+    });
 
     // if root creates a indexable canister, cascade
-    if is_root && is_indexable {
-        SubnetIndex::insert(&canister, new_canister_pid)?;
+    if created_under_root && uses_directory {
+        SubnetDirectory::insert(&canister, new_canister_pid)?;
 
         // update directly as it won't yet be in the child index
         let bundle = StateBundle::subnet_index();