+++ crates/icu/src/macros/endpoints.rs
@@ -62,14 +62,14 @@ macro_rules! icu_endpoints {
 
         #[::icu::ic::query]
         pub fn icrc10_supported_standards() -> Vec<(String, String)> {
-            $crate::state::Icrc10Registry::supported_standards()
+            $crate::state::icrc::Icrc10Registry::supported_standards()
         }
 
         #[::icu::ic::query]
         async fn icrc21_canister_call_consent_message(
             req: ::icu::interface::icrc::Icrc21ConsentMessageRequest,
         ) -> ::icu::interface::icrc::Icrc21ConsentMessageResponse {
-            $crate::state::Icrc21Registry::consent_message(req)
+            $crate::state::icrc::Icrc21Registry::consent_message(req)
         }
 
         //
@@ -111,8 +111,8 @@ macro_rules! icu_endpoints {
         }
 
         #[::icu::ic::query]
-        fn icu_subnet_index() -> ::icu::memory::SubnetIndexView {
-            $crate::memory::SubnetIndex::export()
+        fn icu_subnet_directory() -> ::icu::memory::SubnetDirectoryView {
+            $crate::memory::SubnetDirectory::export()
         }
 
         #[::icu::ic::query]
@@ -127,24 +127,24 @@ macro_rules! icu_endpoints {
         #[::icu::ic::query]
         async fn icu_delegation_get(
             session_pid: Principal,
-        ) -> Result<::icu::state::DelegationSessionView, ::icu::Error> {
-            $crate::state::DelegationRegistry::get(session_pid)
+        ) -> Result<::icu::state::delegation::DelegationSessionView, ::icu::Error> {
+            $crate::state::delegation::DelegationRegistry::get(session_pid)
         }
 
         #[::icu::ic::update]
         async fn icu_delegation_track(
             session_pid: Principal,
-        ) -> Result<::icu::state::DelegationSessionView, ::icu::Error> {
-            $crate::state::DelegationRegistry::track(msg_caller(), session_pid)
+        ) -> Result<::icu::state::delegation::DelegationSessionView, ::icu::Error> {
+            $crate::state::delegation::DelegationRegistry::track(msg_caller(), session_pid)
         }
 
         #[::icu::ic::update]
         async fn icu_delegation_register(
-            args: ::icu::state::RegisterSessionArgs,
+            args: ::icu::state::delegation::RegisterSessionArgs,
         ) -> Result<(), ::icu::Error> {
             $crate::auth_require_any!(::icu::auth::is_whitelisted)?;
 
-            $crate::state::DelegationRegistry::register_session(msg_caller(), args)
+            $crate::state::delegation::DelegationRegistry::register_session(msg_caller(), args)
         }
 
         #[::icu::ic::update]
@@ -156,7 +156,7 @@ macro_rules! icu_endpoints {
             let expected = pid;
             $crate::auth_require_any!(is_parent, |caller| is_principal(caller, expected))?;
 
-            $crate::state::DelegationRegistry::revoke_session_or_wallet(pid)
+            $crate::state::delegation::DelegationRegistry::revoke_session_or_wallet(pid)
         }
     };
 }