+++ crates/icu/src/memory/canister/mod.rs
@@ -1,7 +1,0 @@
-mod canister_state;
-mod child_index;
-mod cycle_tracker;
-
-pub use canister_state::*;
-pub use child_index::*;
-pub use cycle_tracker::*;