+++ crates/icu/src/memory/mod.rs
@@ -1,23 +1,35 @@
-pub mod app;
-pub mod canister;
 pub mod memory_registry;
-pub mod subnet;
 
-pub use app::{AppState, AppStateData};
-pub use canister::{
-    CanisterState, CanisterStateData, ChildIndex, ChildIndexView, CycleTracker, CycleTrackerView,
-};
-pub use memory_registry::{MemoryRegistry, MemoryRegistryError, MemoryRegistryView};
-pub use subnet::{SubnetIndex, SubnetIndexView};
+pub mod cells {
+    mod app_state;
+    mod canister_state;
 
-use crate::{
-    ic::structures::{DefaultMemoryImpl, memory::MemoryManager},
-    memory::{
-        app::AppStateError,
-        canister::{CanisterStateError, ChildIndexError},
-        subnet::SubnetIndexError,
-    },
-};
+    pub use app_state::*;
+    pub use canister_state::*;
+}
+
+pub mod trackers {
+    mod cycle_tracker;
+
+    pub use cycle_tracker::*;
+}
+
+pub mod trees {
+    mod child_index;
+    mod subnet_directory;
+    mod subnet_registry;
+
+    pub use child_index::*;
+    pub use subnet_directory::*;
+    pub use subnet_registry::*;
+}
+
+pub use cells::*;
+pub use memory_registry::*;
+pub use trackers::*;
+pub use trees::*;
+
+use crate::ic::structures::{DefaultMemoryImpl, memory::MemoryManager};
 use std::cell::RefCell;
 use thiserror::Error as ThisError;
 
@@ -29,9 +41,12 @@ pub(crate) const MEMORY_REGISTRY_MEMORY_ID: u8 = 0;
 
 pub(crate) const APP_STATE_MEMORY_ID: u8 = 1;
 pub(crate) const CANISTER_STATE_MEMORY_ID: u8 = 2;
-pub(crate) const CHILD_INDEX_MEMORY_ID: u8 = 3;
-pub(crate) const SUBNET_INDEX_MEMORY_ID: u8 = 4;
-pub(crate) const CYCLE_TRACKER_MEMORY_ID: u8 = 5;
+
+pub(crate) const SUBNET_REGISTRY_MEMORY_ID: u8 = 3;
+pub(crate) const SUBNET_DIRECTORY_MEMORY_ID: u8 = 4;
+pub(crate) const CHILD_INDEX_MEMORY_ID: u8 = 5;
+
+pub(crate) const CYCLE_TRACKER_MEMORY_ID: u8 = 10;
 
 //
 // MEMORY_MANAGER
@@ -66,5 +81,8 @@ pub enum MemoryError {
     ChildIndexError(#[from] ChildIndexError),
 
     #[error(transparent)]
-    SubnetIndexError(#[from] SubnetIndexError),
+    SubnetDirectoryError(#[from] SubnetDirectoryError),
+
+    #[error(transparent)]
+    SubnetRegistryError(#[from] SubnetRegistryError),
 }