+++ crates/icu/src/memory/subnet/mod.rs
@@ -1,3 +1,0 @@
-mod subnet_index;
-
-pub use subnet_index::*;